+++ www/go/modules/community/tasks/views/extjs3/ChooseTasklistDialog.js
@@ -1,12 +1,28 @@
-go.modules.community.tasks.ChooseTasklistDialog = Ext.extend(Ext.Window, {
-	title: t("Choose a task list"),
+go.modules.community.tasks.ChooseTasklistDialog = Ext.extend(go.Window, {
+	title: t("Choose a tasklist"),
     entityStore: "Task",
-    layout: 'fit',
-	width: dp(800),
-	height: dp(800),
+    layout: 'form',
+  width: dp(800),
+  height: dp(800),
+    modal: true,
 
 	initComponent: function () {
-        this.chooseTasklistGrid = new go.modules.community.tasks.ChooseTasklistGrid();
+        this.chooseTasklistGrid = new go.modules.community.tasks.ChooseTasklistGrid({
+            height: dp(700)
+        });
+
+        this.taskListFromCsvCB = new Ext.form.Checkbox({
+            xtype: 'xcheckbox',
+            boxLabel: t('Import task list ID from CSV file'),
+            handler: (cb,checked) => {
+                const el = this.chooseTasklistGrid.getEl();
+                if(checked) {
+                    el.mask()
+                } else {
+                    el.unmask();
+                }
+            }
+        });
 
         this.openFileButton = new Ext.Button({
             iconCls: 'ic-search',
@@ -14,7 +30,7 @@ go.modules.community.tasks.ChooseTasklistDialog = Ext.extend(Ext.Window, {
             width: dp(40),
             height: dp(30),
             handler: function() {
-                if(!this.chooseTasklistGrid.selectedId) {
+                if(!this.chooseTasklistGrid.selectedId && !this.taskListFromCsvCB.checked) {
                     Ext.Msg.show({
                         title:t("Task list not selected"),
                         msg: t("You have not selected any task list. Select a task list before proceeding."),
@@ -23,10 +39,15 @@ go.modules.community.tasks.ChooseTasklistDialog = Ext.extend(Ext.Window, {
                         icon: Ext.MessageBox.WARNING
                      });
                 } else {
+                    let TLvalues = {};
+                    if(!this.taskListFromCsvCB.checked) {
+                        TLvalues = {tasklistId: this.chooseTasklistGrid.selectedId};
+                    }
                     go.util.importFile(
                         'Task', 
                         ".ics,.csv",
-                        { tasklistId: this.chooseTasklistGrid.selectedId },
+                        // { tasklistId: this.chooseTasklistGrid.selectedId },
+                        TLvalues,
                         {},
                         {
                             labels: {
@@ -47,9 +68,29 @@ go.modules.community.tasks.ChooseTasklistDialog = Ext.extend(Ext.Window, {
             scope: this
         });
 
+        const propertiesPanel = new Ext.Panel({
+            hideMode : 'offsets',
+            //title : t("Properties"),
+            labelAlign: 'top',
+            layout : 'form',
+            autoScroll : true,
+            items : [{
+                xtype: "container",
+                layout: "form",
+                defaults: {
+                    anchor: '100%'
+                },
+                labelWidth: 16,
+                items: [
+                    this.taskListFromCsvCB,
+                    this.chooseTasklistGrid
+                ]
+            }]
+        });
+
         this.buttons = [this.openFileButton];
 
-		this.items = [
+		    this.items = [
             this.chooseTasklistGrid
         ];