+++ www/go/modules/community/tasks/views/extjs3/ChooseTasklistGrid.js
@@ -46,8 +46,8 @@ go.modules.community.tasks.ChooseTasklistGrid = Ext.extend(go.grid.GridPanel, {
             listeners: {
 				scope: this,
 				rowclick: function (grid, rowIndex, e) {
-                    var row = this.getSelectionModel().getSelections()[0];
-                    this.selectedId = row.get("id");
+					const row = this.getSelectionModel().getSelections()[0];
+					this.selectedId = row.get("id");
 				}
 			},
 			autoExpandColumn: 'name',